updates typing and parsing to include confirmations - adds new components to Proposal Details (safe)
