+++ src/components/Proposals/MultisigDetails/TxActions.tsx
@@ -1,0 +1,9 @@
+import ContentBox from '../../ui/ContentBox';
+
+export function TxActions() {
+  return (
+    <ContentBox bg="black.900-semi-transparent">
+      <div></div>
+    </ContentBox>
+  );
+}
